requirements.txt
python-dotenv
numpy
pandas
pymongo
certifi
pymongo[srv]
scikit-learn

##-e .
